refactor(server): extract /api/social-generator/* into a router module
Third route-group extraction. Moves all 6 social-generator handlers +
the social-only ensureSocialPostsTable (table init) out of server.js into
src/server/routes/social-generator.js, mounted via
`app.use('/api/social-generator', requireAuth, socialGeneratorRouter)`.

Routes (paths byte-identical, guard-verified):
  GET  /api/social-generator/generate           (SSE)
  GET  /api/social-generator/arcs/:brandProfileId
  POST /api/social-generator/regenerate-arcs/:brandProfileId
  GET  /api/social-generator/recent/:brandProfileId
  POST /api/social-generator/edit/:postId
  POST /api/social-generator/publish-x/:postId

Auth: requireAuth at the mount (all 6 uniformly authed).

ensureSocialPostsTable (generated_social_posts schema init) + its
boot-time call moved into the module — the init now fires when server.js
imports the router, preserving the startup behavior.

Dependencies — all already extracted: pool (db), anthropic + dateContext
(llm), safeParseLLM (llm-json), verifyBrandAccess (auth), activeStreams
(streams), buildSocialImagePrompt + generateSocialImage (images),
buildXOAuthHeader + uploadXMedia + refreshXOAuth2Token (x) — the publish-x
handler exercises the x.js helpers. Plus fs/path/randomUUID (Node).
server.js no longer references activeStreams (SG was its last user), so
that import is dropped.

Handler bodies verbatim; only registration lines changed.

Test plan:
- node --check server.js + module -> OK
- npm run lint (no-undef) -> clean (it caught fs/path/randomUUID on the
  first pass — added them; that's the gate doing its job)
- route-inventory guard -> 213 byte-identical (6 SG paths reconstructed)
- vitest -> 93/93

Rollback: revert; verbatim move, restores inline + drops mount.
